Fixed collapse/expand of event
